Codechange: split "main.py" into several files
This makes it a bit easier to understand what belongs where, and
makes a bit easier to see what happens in main.py.
