+++ webclient/helpers.py
@@ -1,0 +1,91 @@
+import flask
+import requests
+import urllib
+
+from .main import app
+
+_api_url = "http://localhost:8080"  # TODO environment or something
+_frontend_url = "https://localhost:5000"  # TODO
+
+
+def template(*args, **kwargs):
+    if "message" in kwargs:
+        kwargs.setdefault("messages", []).append(kwargs["message"])
+
+    response = flask.make_response(flask.render_template(*args, **kwargs))
+    response.headers["Content-Security-Policy"] = "default-src 'self'"
+    return response
+
+
+def external_url_for(*args, **kwargs):
+    return _frontend_url + flask.url_for(*args, **kwargs)
+
+
+def redirect(*args, **kwargs):
+    return flask.redirect(flask.url_for(*args, **kwargs))
+
+
+def not_found():
+    flask.abort(404)
+
+
+def api_error():
+    flask.abort(500)
+
+
+def api_call(method, path, params=None, json=None, session=None, return_errors=False):
+    url = _api_url + "/" + "/".join(urllib.parse.quote(p, safe="") for p in path)
+    headers = None
+    if session and session.api_token:
+        headers = {"Authorization": "Bearer " + session.api_token}
+    try:
+        app.logger.info("API request to '{}': {}".format(url, json))
+        r = method(url, params=params, headers=headers, json=json)
+
+        success = r.status_code in (200, 201, 204)
+        if not success:
+            app.logger.warning("API failed: {}".format(r.text))
+
+        if success:
+            result = None
+            try:
+                result = r.json()
+            except Exception:
+                result = None
+            if return_errors:
+                return (result, None)
+            else:
+                return result
+        elif return_errors:
+            error = str(r.json().get("errors", "API call failed"))
+            return (None, error)
+        elif r.status_code == 404:
+            not_found()
+        elif r.status_code == 401:
+            if session and session.is_auth:
+                redirect("root", message="Access denied")
+            else:
+                redirect("login")
+    except Exception:
+        pass
+
+    if return_errors:
+        return (None, "API call failed")
+    else:
+        api_error()
+
+
+def api_get(*args, **kwargs):
+    return api_call(requests.get, *args, **kwargs)
+
+
+def api_post(*args, **kwargs):
+    return api_call(requests.post, *args, **kwargs)
+
+
+def api_put(*args, **kwargs):
+    return api_call(requests.put, *args, **kwargs)
+
+
+def api_delete(*args, **kwargs):
+    return api_call(requests.delete, *args, **kwargs)
